Updated DataGrid props to prevent issue with column header misalignment.
Based off this issue https://github.com/mui/mui-x/pull/2673#issue-1004950963, I believe the reason for the column issue has to do with how mui is virtualizing the headers. I increased the columnBuffer to a large enough number so all the column headers are always generated. Prior to this, I could regularly break the headers, but I haven't been able to since.

In theory, this is slower, but it's a very small amount of content and the trade-off seems well worth it, because there isn't really a way to solve the virtualization issue if my diagnosis is correct.